% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data}
\alias{data}
\title{Combined forecast data for the paper}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 61296 rows and 12 columns.
}
\usage{
data
}
\description{
Combined forecast data for the paper
}
\keyword{datasets}
