% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_forecast_date}
\alias{get_forecast_date}
\title{Get Target Forecast Date}
\usage{
get_forecast_date(dir, char = FALSE)
}
\arguments{
\item{dir}{A character string indicating the path to the target data folder.}

\item{char}{Logical, defaults to \code{TRUE}. Should the date be
returned as a character string}
}
\value{
A date or a character string representing a date.
}
\description{
Get Target Forecast Date
}
