% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{log_data}
\alias{log_data}
\title{Combined forecast data for the paper on the log scale}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 61296 rows and 13 columns.
}
\usage{
log_data
}
\description{
Combined forecast data for the paper on the log scale
}
\keyword{datasets}
