% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{models}
\alias{models}
\title{Names of the different models and forecasters}
\format{
An object of class \code{list} of length 2.
}
\usage{
models
}
\description{
Names of the different models and forecasters
}
\keyword{datasets}
