% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{predictions}
\alias{predictions}
\title{Forecast data used for the paper}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 74060 rows and 10 columns.
}
\usage{
predictions
}
\description{
Forecast data used for the paper
}
\keyword{datasets}
