% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{study_dates}
\alias{study_dates}
\title{Relevant study dates}
\format{
An object of class \code{list} of length 4.
}
\usage{
study_dates
}
\description{
Relevant study dates
}
\keyword{datasets}
