% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{truth}
\alias{truth}
\title{Truth data used for the paper}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 174 rows and 6 columns.
}
\usage{
truth
}
\description{
Truth data used for the paper
}
\keyword{datasets}
